requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
clang==14.0
cloudpickle==2.2.0
colorama==0.4.5
contourpy==1.0.5
cramjam==2.5.0
cycler==0.11.0
dill==0.3.6
etils==0.8.0
fastparquet==0.8.3
flatbuffers==2.0.7
fonttools==4.37.4
fsspec==2022.8.2
funcsigs==1.0.2
gast==0.4.0
gin-config==0.5.0
google-auth-oauthlib==0.4.6
google-auth==2.12.0
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
gym-notices==0.0.8
gym==0.26.2
h5py==3.7.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
jax==0.3.15
jaxlib==0.3.14
joblib==1.2.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
matplotlib==3.6.1
numpy==1.23.5
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
pip==23.0.0
promise==2.3
progressbar2==4.2.0
protobuf==3.19.6
psutil==5.9.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
sqlalchemy==1.3.24
scikit-learn==1.1.3
scipy==1.9.3
setuptools==65.4.1
six==1.16.0
termcolor==2.0.1
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.1
trax==1.4.1
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.9.0
keras~=2.10.0
tensorflow~=2.10.0